requirements.txt
# HealthGuard IA — Dépendances Python
# Généré le 2025-01-26 — Python 3.13+

# === API & Web ===
fastapi==0.136.1
uvicorn[standard]==0.34.3
pydantic==2.12.5
httpx==0.28.1

# === Machine Learning ===
xgboost==3.2.0
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.0

# === Sécurité ===
argon2-cffi==21.3.0
cryptography>=42.0.0
pyOpenSSL>=24.0.0

# === Base de données ===
# sqlite3 est inclus dans Python standard
psycopg2-binary>=2.9.0
aiofiles>=23.0.0

# === Tests ===
pytest==9.0.3
pytest-cov==7.1.0
pytest-anyio==0.0.0
anyio==4.12.1
httpx==0.28.1  # Pour TestClient FastAPI

# === Utilitaires ===
python-dotenv>=1.0.0

# === TFLite (optionnel — pour conversion modèle) ===
# tensorflow>=2.16.0  # Décommenter pour conversion TFLite
# tflite-support>=0.4.4  # Décommenter pour inférence TFLite sur Android

# === Notes d'installation ===
# pip install -r requirements.txt
# Pour Android : utiliser tflite-runtime au lieu de tensorflow complet
# Le modèle ML est pré-entraîné et fourni sous src/ml/healthguard_model.pkl
